fix refactored utility functions from historyRepository to History Transoformers
